Extract code into `embed`
