+++ src/language-html/embed/class-names.js
@@ -1,0 +1,5 @@
+function printClassNames(value) {
+  return value.trim().split(/\s+/).join(" ");
+}
+
+export default printClassNames;
