+++ src/language-html/embed/style.js
@@ -1,0 +1,8 @@
+function printStyleAttribute(value, attributeTextToDoc) {
+  return attributeTextToDoc(value, {
+    parser: "css",
+    __isHTMLStyleAttribute: true,
+  });
+}
+
+export { printStyleAttribute };
